requirements.txt
transformers==4.49.0
bitsandbytes
peft==0.14.0
pillow==11.1.0
torch>=2.4.1
runpod
maestro[qwen_2_5_vl]==1.1.0rc2
qwen-vl-utils==0.0.10